Create Countries
Using one of the data sets provided by Khan Academy, specifically the Countries and Dependent Territories (2020) data set, I demonstrate an advanced-level understanding of SELECT queries to discover things about the data.  This data set explores the MIN, MAX, and AVG values of the data and filters it using HAVING, AND/OR, and CASE to interpret and and better understand relationships within the data.